backend/requirements.txt
fastapi==0.101.0
uvicorn[standard]==0.23.2
asyncer==0.0.2
python-multipart==0.0.6
deta==1.2.0
feedparser==6.0.10
slack_sdk==3.21.3
aiohttp==3.8.5
python-slugify==8.0.1
